remove logic from components, add new gen container, types componets props
