--- src/view/components/MainKeyboard/index.tsx
@@ -1,164 +1,0 @@
-// Core
-import React, { FC } from 'react';
-
-// Bus
-import { useTogglersRedux } from '../../../bus/client/togglers';
-
-// Styles
-import * as S from './styles';
-
-// Components
-import { Keyboard } from '../';
-
-// Types
-type PropTypes = {
-    keyboardRef: React.MutableRefObject<HTMLDivElement | null>
-}
-
-export const MainKeyboard: FC<PropTypes> = ({ keyboardRef }) => {
-    const { togglersRedux } = useTogglersRedux();
-
-    const RuLayout = {
-        arrayElemsNumber: [
-            { value: '1', keycode: 49 },
-            { value: '2', keycode: 50 },
-            { value: '3', keycode: 51 },
-            { value: '4', keycode: 52 },
-            { value: '5', keycode: 53 },
-            { value: '6', keycode: 54 },
-            { value: '7', keycode: 55 },
-            { value: '8', keycode: 56 },
-            { value: '9', keycode: 57 },
-            { value: '0', keycode: 48 },
-        ],
-
-        arrayElemsWord: [
-            { value: 'й', keycode: 81 },
-            { value: 'ц', keycode: 87 },
-            { value: 'у', keycode: 69 },
-            { value: 'к', keycode: 82 },
-            { value: 'е', keycode: 84 },
-            { value: 'н', keycode: 89 },
-            { value: 'г', keycode: 85 },
-            { value: 'ш', keycode: 73 },
-            { value: 'щ', keycode: 79 },
-            { value: 'з', keycode: 80 },
-            { value: 'х', keycode: 219 },
-        ],
-
-        arrayWordGrid: [
-            { value: 'ф', keycode: 65 },
-            { value: 'ы', keycode: 83 },
-            { value: 'в', keycode: 68 },
-            { value: 'а', keycode: 70 },
-            { value: 'п', keycode: 71 },
-            { value: 'р', keycode: 72 },
-            { value: 'о', keycode: 74 },
-            { value: 'л', keycode: 75 },
-            { value: 'д', keycode: 76 },
-            { value: 'ж', keycode: 186 },
-            { value: 'э', keycode: 222 },
-        ],
-
-        arrayElemsWordTwo: [
-            { value: 'Shift', keycode: 16 },
-            { value: 'я', keycode: 90 },
-            { value: 'ч', keycode: 88 },
-            { value: 'с', keycode: 67 },
-            { value: 'м', keycode: 86 },
-            { value: 'и', keycode: 66 },
-            { value: 'т', keycode: 78 },
-            { value: 'ь', keycode: 77 },
-            { value: 'б', keycode: 188 },
-            { value: 'ю', keycode: 190 },
-            { value: 'Backspace', keycode: 8 },
-        ],
-
-        arrayElemsWordThree: [
-            { value: ',', keycode: 188 },
-            { value: 'Ru', keycode: 19 },
-            { value: 'Пробел', keycode: 32 },
-            { value: '.', keycode: 190 },
-            { value: 'Enter', keycode: 13 },
-        ],
-    };
-
-    const EnLayout = {
-        arrayElemsNumber: [
-            { value: '1', keycode: 49 },
-            { value: '2', keycode: 50 },
-            { value: '3', keycode: 51 },
-            { value: '4', keycode: 52 },
-            { value: '5', keycode: 53 },
-            { value: '6', keycode: 54 },
-            { value: '7', keycode: 55 },
-            { value: '8', keycode: 56 },
-            { value: '9', keycode: 57 },
-            { value: '0', keycode: 48 },
-        ],
-
-        arrayElemsWord: [
-            { value: 'q', keycode: 81 },
-            { value: 'w', keycode: 87 },
-            { value: 'e', keycode: 69 },
-            { value: 'r', keycode: 82 },
-            { value: 't', keycode: 84 },
-            { value: 'y', keycode: 89 },
-            { value: 'u', keycode: 85 },
-            { value: 'i', keycode: 73 },
-            { value: 'o', keycode: 79 },
-            { value: 'p', keycode: 80 },
-        ],
-
-        arrayWordGrid: [
-            { value: 'a', keycode: 65 },
-            { value: 's', keycode: 83 },
-            { value: 'd', keycode: 68 },
-            { value: 'f', keycode: 70 },
-            { value: 'g', keycode: 71 },
-            { value: 'h', keycode: 72 },
-            { value: 'j', keycode: 74 },
-            { value: 'k', keycode: 75 },
-            { value: 'l', keycode: 76 },
-        ],
-
-        arrayElemsWordTwo: [
-            { value: 'Shift', keycode: 16 },
-            { value: 'z', keycode: 90 },
-            { value: 'x', keycode: 88 },
-            { value: 'c', keycode: 67 },
-            { value: 'v', keycode: 86 },
-            { value: 'b', keycode: 66 },
-            { value: 'n', keycode: 78 },
-            { value: 'm', keycode: 77 },
-            { value: 'Backspace', keycode: 8 },
-        ],
-
-        arrayElemsWordThree: [
-            { value: ',', keycode: 188 },
-            { value: 'En', keycode: 19 },
-            { value: 'Space', keycode: 32 },
-            { value: '.', keycode: 190 },
-            { value: 'Enter', keycode: 13 },
-        ],
-    };
-
-    return (
-        <S.Container>
-            {
-                togglersRedux.isRuLayout
-                    ? (
-                        <Keyboard
-                            arrayKeyboardWords = { RuLayout }
-                            keybortRef = { keyboardRef }
-                        />
-                    ) : (
-                        <Keyboard
-                            arrayKeyboardWords = { EnLayout }
-                            keybortRef = { keyboardRef }
-                        />
-                    )
-            }
-        </S.Container>
-    );
-};
