--- src/view/components/MainKeyboard/styles.ts
@@ -1,5 +1,0 @@
-// Core
-import styled from 'styled-components';
-
-export const Container = styled.section`
-`;
