+++ src/view/components/MessageInfoActions/index.tsx
@@ -1,10 +1,6 @@
 // Core
 import React, { FC } from 'react';
 
-// Bus
-import { useSelectedMessage } from '../../../bus/client/selectedMessage';
-import { useMessages } from '../../../bus/messages';
-
 // Styles
 import * as S from './styles';
 
@@ -12,43 +8,36 @@ import * as S from './styles';
 import { Message } from '../../../bus/messages/types';
 
 type PropTypes = {
+    isEditingMessage: boolean
     message: Message
+    closeSelectedMessage: () => void
+    changeSelectedMessage: (message: Message) => void
+    handleRemoveMessage: (message: Message) => void
 }
 
-export const MessageInfoActions: FC<PropTypes> = ({ message }) => {
-    const {
-        deleteMessage,
-    } = useMessages();
-
-    const {
-        changeSelectedMessage, selectedMessage, closeSelectedMessage,
-    } = useSelectedMessage();
-
-    return (
-        <S.Container>
-            <S.BtnsBox>
-                {
-                    selectedMessage?._id === message._id ? (
-                        <S.CloseBtn onClick = { () => void closeSelectedMessage() }>
-                            Close
-                        </S.CloseBtn>
-                    ) : (
-                        <S.BtnChangeMessage onClick = { () => void changeSelectedMessage(message) }>
-                            change
-                        </S.BtnChangeMessage>
-                    )
-                }
-                <S.BtnRemoveMessage onClick = { () => {
-                    // eslint-disable-next-line no-alert
-                    const isDelete = confirm('do you really want to delete messages');
-
-                    if (isDelete) {
-                        deleteMessage(message);
+export const MessageInfoActions: FC<PropTypes>
+    = ({
+        changeSelectedMessage, closeSelectedMessage, handleRemoveMessage,
+        isEditingMessage, message,
+    }) => {
+        return (
+            <S.Container>
+                <S.BtnsBox>
+                    {
+                        isEditingMessage ? (
+                            <S.CloseBtn onClick = { closeSelectedMessage }>
+                                Close
+                            </S.CloseBtn>
+                        ) : (
+                            <S.BtnChangeMessage onClick = { () => changeSelectedMessage(message) }>
+                                change
+                            </S.BtnChangeMessage>
+                        )
                     }
-                }  }>
-                    remove
-                </S.BtnRemoveMessage>
-            </S.BtnsBox>
-        </S.Container>
-    );
-};
+                    <S.BtnRemoveMessage onClick = { () => handleRemoveMessage(message) }>
+                        remove
+                    </S.BtnRemoveMessage>
+                </S.BtnsBox>
+            </S.Container>
+        );
+    };
